Styled experience section
