+++ src/components/Header/index.js
@@ -4,7 +4,7 @@ const navItems = [
   { title: 'Home', label: 'home', url: '/#' },
   { title: 'About', label: 'about-me', url: '/#about-me' },
   { title: 'Projects', label: 'my-projects', url: '/#my-projects' },
-  { title: 'Experience', label: 'experience', url: '/#experience' },
+  { title: 'Journey', label: 'journey', url: '/#journey' },
   { title: 'Skills', label: 'skills', url: '/#skills' }
 ]
