+++ src/components/Projects/projects.module.css
@@ -13,7 +13,7 @@
   justify-content: center;
   text-decoration: none;
   gap: 8px;
-  margin: 10px 0;
+  marginBottom: 10px;
 }
