+++ src/components/Section/SectionTitle.js
@@ -10,7 +10,7 @@ const styles = {
   h2: {
     alignSelf: 'flex-start',
     marginTop: '2.75rem',
-    marginBottom: '0.75rem',
+    marginBottom: '1rem',
     display: 'flex',
     justifyContent: 'center',
     gap: '8px'
